feat(device-hub): unify virtual device management
