+++ src/renderer/components/controllersHub/DeviceOverviewPanel.vue
@@ -1,0 +1,89 @@
+<template>
+  <section class="chub-panel">
+    <div class="chub-window" :class="runtimeTone">
+      <span class="chub-window-tab">{{ t.deviceHub.runtime.title }}</span>
+      <div class="chub-hud-row">
+        <div class="chub-hud-state"><span class="chub-status-dot"></span><strong>{{ runtimeLabel }}</strong></div>
+        <el-button size="small" :loading="loading" @click="refresh">{{ t.deviceHub.refresh }}</el-button>
+      </div>
+      <div class="chub-runtime-grid">
+        <div><span>{{ t.deviceHub.runtime.host }}</span><strong>{{ hostLabel }}</strong></div>
+        <div><span>{{ t.deviceHub.runtime.transport }}</span><strong>{{ transportLabel }}</strong></div>
+        <div><span>{{ t.deviceHub.runtime.owner }}</span><strong>Sunshine Core</strong></div>
+      </div>
+      <p v-if="loadError" class="chub-status-error">{{ t.deviceHub.statusUnavailable }}</p>
+    </div>
+
+    <div class="chub-cards chub-overview-cards">
+      <article class="chub-card">
+        <div class="chub-card-head"><strong>{{ t.deviceHub.overview.controllers }}</strong><el-tag size="small" :type="controllerReady ? 'success' : 'info'" effect="plain">{{ controllerReady ? t.deviceHub.available : t.deviceHub.needsSetup }}</el-tag></div>
+        <p class="chub-hint">{{ t.deviceHub.overview.controllersHint }}</p>
+        <div class="chub-card-actions"><el-button size="small" type="primary" @click="emit('navigate', 'controllers')">{{ t.deviceHub.overview.manageControllers }}</el-button></div>
+      </article>
+      <article class="chub-card">
+        <div class="chub-card-head"><strong>{{ t.deviceHub.overview.microphone }}</strong><el-tag size="small" :type="microphoneTagType" effect="plain">{{ microphoneTag }}</el-tag></div>
+        <p class="chub-hint">{{ microphoneSummary }}</p>
+        <div class="chub-card-actions"><el-button size="small" type="primary" @click="emit('navigate', 'microphone')">{{ t.deviceHub.overview.inspectMicrophone }}</el-button></div>
+      </article>
+      <article class="chub-card">
+        <div class="chub-card-head"><strong>{{ t.deviceHub.overview.usb }}</strong><el-tag size="small" type="info" effect="plain">{{ t.deviceHub.planned }}</el-tag></div>
+        <p class="chub-hint">{{ t.deviceHub.overview.usbHint }}</p>
+        <div class="chub-card-actions"><el-button size="small" @click="emit('navigate', 'usb')">{{ t.deviceHub.overview.inspectUsb }}</el-button></div>
+      </article>
+    </div>
+  </section>
+</template>
+
+<script setup>
+import { computed, onMounted, reactive, ref } from 'vue'
+import { dualsense, virtualMicrophone } from '../../tauri-adapter.js'
+import { deviceRuntimeReady, microphoneOverviewState } from '../../composables/deviceHubStatus.js'
+import { useI18n } from '../../desktop/i18n/index.js'
+
+const emit = defineEmits(['navigate'])
+const { t } = useI18n()
+const loading = ref(false)
+const loadError = ref(false)
+const ds = reactive({ verified: false, usbip_available: false, usbip_version: '' })
+const mic = reactive({ component_available: false, device_created: false, host_streaming: false })
+const controllerReady = computed(() => ds.verified)
+const microphoneState = computed(() => microphoneOverviewState(mic))
+const microphoneTag = computed(() => ({
+  missing: t.value.deviceHub.overview.statusUnavailable,
+  capturing: t.value.deviceHub.overview.statusCapturing,
+  idle: t.value.deviceHub.overview.statusReady,
+  waiting: t.value.deviceHub.overview.statusWaiting,
+})[microphoneState.value])
+const microphoneTagType = computed(() => ({
+  missing: 'info',
+  capturing: 'primary',
+  idle: 'success',
+  waiting: 'info',
+})[microphoneState.value])
+const runtimeReady = computed(() => deviceRuntimeReady(ds, mic))
+const runtimeTone = computed(() => runtimeReady.value ? 'state-ready' : loadError.value ? 'state-error' : '')
+const runtimeLabel = computed(() => runtimeReady.value ? t.value.deviceHub.runtime.ready : t.value.deviceHub.runtime.needsAttention)
+const hostLabel = computed(() => runtimeReady.value ? t.value.deviceHub.available : t.value.deviceHub.unavailable)
+const transportLabel = computed(() => ds.usbip_available ? `USB/IP ${ds.usbip_version || ''}`.trim() : t.value.deviceHub.unavailable)
+const microphoneSummary = computed(() => {
+  const key = microphoneState.value
+  return {
+    missing: t.value.deviceHub.overview.microphoneMissing,
+    capturing: t.value.deviceHub.overview.microphoneCapturing,
+    idle: t.value.deviceHub.overview.microphoneIdle,
+    waiting: t.value.deviceHub.overview.microphoneWaiting,
+  }[key]
+})
+
+async function refresh() {
+  if (loading.value) return
+  loading.value = true
+  loadError.value = false
+  const [dsResult, micResult] = await Promise.all([dualsense.getStatus(), virtualMicrophone.getStatus()])
+  if (dsResult?.success) Object.assign(ds, dsResult.data)
+  if (micResult?.success) Object.assign(mic, micResult.data)
+  loadError.value = !dsResult?.success && !micResult?.success
+  loading.value = false
+}
+onMounted(refresh)
+</script>
