+++ src/renderer/desktop/i18n/en.js
@@ -960,7 +960,9 @@ Output only one short sentence, no explanation. Reply in English.`,
     virtualDisplay: 'Virtual Display',
     webStream: 'Web Stream',
     aiAssistant: 'Mita',
+    deviceHub: 'Device Hub',
     controllers: 'Controllers',
+    controllersHub: 'Controllers Hub',
     dualSense: 'DualSense Emulation',
     uninstallVdd: 'Uninstall VDD',
     restartGpu: 'Restart GPU Driver',
@@ -981,6 +983,118 @@ Output only one short sentence, no explanation. Reply in English.`,
     eyebrow: 'MANAGE / CONTROLLERS',
     title: 'Controllers',
     intro: 'Manage Sunshine virtual controllers.',
+  },
+
+  deviceHub: {
+    eyebrow: 'MANAGE / DEVICE HUB',
+    title: 'Device Hub',
+    intro: 'Manage virtual devices, local components, and the boundary for future USB passthrough. Streaming settings remain in the WebUI.',
+    refresh: 'Refresh',
+    available: 'Available',
+    unavailable: 'Unavailable',
+    needsSetup: 'Needs setup',
+    experimental: 'Experimental',
+    planned: 'Planned',
+    unknown: 'Unknown',
+    statusUnavailable: 'Device status is unavailable. Make sure Sunshine is running and up to date.',
+    controllerAdvanced: 'Advanced controller behavior (DS4 / motion / DSU)',
+    tabs: {
+      overview: 'Overview',
+      controllers: 'Controllers',
+      microphone: 'Microphone',
+      usb: 'USB passthrough',
+      components: 'Components & diagnostics',
+    },
+    runtime: {
+      title: 'DEVICE RUNTIME',
+      ready: 'Device runtime is ready',
+      needsAttention: 'Device runtime needs attention',
+      host: 'Virtual Device Host',
+      transport: 'USB/IP Transport',
+      owner: 'Runtime owner',
+    },
+    overview: {
+      controllers: 'Virtual controllers',
+      controllersHint: 'Choose the controller identity games see and manage the optional DualSense component.',
+      manageControllers: 'Manage controllers',
+      microphone: 'Virtual microphone',
+      microphoneMissing: 'The signed Virtual Device Host component is unavailable.',
+      microphoneCapturing: 'A Windows application is currently capturing the virtual endpoint.',
+      microphoneIdle: 'The endpoint is ready and waiting for an application.',
+      microphoneWaiting: 'The component is ready; the endpoint will be created when required.',
+      statusCapturing: 'Capturing',
+      statusReady: 'Ready',
+      statusWaiting: 'Not created',
+      statusUnavailable: 'Unavailable',
+      inspectMicrophone: 'View status & test',
+      usb: 'USB device passthrough',
+      usbHint: 'Future client-to-host devices will be explicitly authorized and bound to one stream.',
+      inspectUsb: 'View planned boundary',
+    },
+    microphone: {
+      title: 'VIRTUAL MICROPHONE',
+      experimentalTitle: 'Experimental USB/IP microphone',
+      experimentalHint: 'This path uses the signed Virtual Device Host and pinned USB/IP transport. It does not use VB-Cable loopback.',
+      test: 'Test microphone path',
+      testSuccess: 'The test tone was injected into the virtual microphone.',
+      testFailed: 'The microphone path test did not complete.',
+      configuredBackend: 'Configured backend',
+      activeBackend: 'Active backend',
+      component: 'Host component',
+      endpoint: 'Windows endpoint',
+      capture: 'Application capture',
+      buffer: 'Host buffer',
+      health: 'Transport health',
+      healthOk: 'Healthy',
+      healthAttention: 'Needs attention',
+      endpointReady: 'Ready',
+      endpointNotCreated: 'Not created',
+      captureActive: 'Capturing',
+      captureIdle: 'Waiting for an app',
+      diagnostics: 'View cumulative diagnostics',
+      underruns: 'Buffer underruns (lifetime)',
+      droppedFrames: 'Unconsumed frames (lifetime)',
+      submitErrors: 'Submit errors (lifetime)',
+      backends: {
+        usbip_experimental: 'USB/IP virtual microphone',
+        auto: 'Automatic',
+        vb_cable: 'VB-Cable compatibility mode',
+        disabled: 'Disabled',
+      },
+      fallback: 'Backend fallback is active',
+      errorHint: 'Review the Sunshine logs and repair the component if the error persists.',
+      webuiHint: 'Enable the microphone and choose its backend in Audio / Video settings.',
+      openWebui: 'Open WebUI audio settings',
+      openFailed: 'Could not open the Sunshine WebUI.',
+      states: {
+        absent: 'Not created',
+        starting: 'Starting',
+        ready: 'Ready',
+        idle: 'Ready',
+        remote_active: 'Receiving remote audio',
+        host_capturing: 'Application is capturing',
+        faulted: 'Faulted',
+        unsupported: 'Unsupported on this platform',
+      },
+    },
+    usb: {
+      title: 'General USB passthrough',
+      notAvailable: 'Not available in this build',
+      hint: 'This section reserves the product boundary without presenting fake devices or controls before the protocol and client support exist.',
+      direction: 'Direction',
+      clientToHost: 'Moonlight client → Sunshine host',
+      ownership: 'Ownership',
+      sessionBound: 'Bound to one authenticated stream',
+      release: 'Disconnect behavior',
+      autoRelease: 'Automatically returned to the client',
+      boundaryTitle: 'Separate from virtual-device creation',
+      boundaryHint: 'Virtual Device Host creates Sunshine-owned devices. A future USB Passthrough Manager will enumerate, authorize, claim, and release external physical devices.',
+    },
+    components: {
+      runtimeTitle: 'Shared runtime components',
+      hostHint: 'Creates Sunshine-owned DualSense and virtual microphone devices.',
+      transportHint: 'Pinned system transport shared by composite virtual devices and future passthrough.',
+    },
   },
 
   dualSense: {
@@ -1122,6 +1236,74 @@ Output only one short sentence, no explanation. Reply in English.`,
           fallback: 'The device test could not be completed. Select “Repair component”, then try again.',
         },
         uninstall: { fallback: 'The DualSense component could not be uninstalled. End any active streams, approve the administrator prompt, and try again.' },
+      },
+    },
+  },
+
+  controllersHub: {
+    eyebrow: 'MANAGE / CONTROLLERS HUB',
+    title: 'Controllers Hub',
+    intro: 'Pick which controller games see while streaming; everything else lives under Advanced.',
+    advancedTitle: 'Advanced (DualSense / motion / drivers & tools)',
+    steamNote: 'Steam Controller support is in development',
+    emulation: {
+      typeQuestion: 'Which controller should games see while streaming?',
+      familyAuto: 'Auto (recommended)',
+      familyAutoHint: 'Best compatibility by default when you are not sure what to choose.',
+      xbox360: 'Xbox 360',
+      xbox360Hint: 'Works with most Windows games when broad compatibility matters most.',
+      ds4ModeHint: 'For games that need PlayStation button layouts, touchpad, or motion input.',
+      ds5Variant: 'DualSense',
+      ds5ModeHint: 'Adds adaptive triggers and HD Haptics; requires the Virtual Device Host component.',
+      sunshineUnreachable: 'Cannot read Sunshine config. Make sure Sunshine is running.',
+      retry: 'Retry',
+      saveFailed: 'Save failed; the previous setting was restored.',
+      perAppTitle: 'Per-game settings',
+      perAppHint: 'Pick a controller type for a single game; it overrides the global choice above.',
+      noApps: 'No apps yet. Add them in the Sunshine WebUI first.',
+      appLoadFailed: 'Failed to load the app list.',
+      appSaveConflict: 'App config was modified elsewhere; the list has been reloaded. Please retry.',
+      appSaved: 'Game controller type saved.',
+      appSaveFailed: 'Failed to save; the previous value was restored.',
+      followGlobal: 'Follow global',
+      ds4Title: 'DualShock 4 details',
+      ds4Hint: 'Only affects input while identified as DualShock 4.',
+      backAsTouchpadClick: 'Map BACK as touchpad click',
+      motionAsDs4: 'Inject motion data as DS4',
+      touchpadAsDs4: 'Inject touchpad as DS4',
+      dsuTitle: 'DSU motion server',
+      dsuHint: 'Lets motion-assist tools read controller pose over UDP.',
+      dsuEnable: 'Enable motion server',
+      dsuPort: 'Port',
+    },
+    hardware: {
+      comingSoonTitle: 'Coming Soon',
+      comingSoonHint: 'This controller type is planned; its dedicated settings are under development.',
+    },
+    peripherals: {
+      title: 'Peripheral Tools',
+      installed: 'Installed',
+      notInstalled: 'Not installed',
+      install: 'Install',
+      uninstall: 'Uninstall',
+      confirmTitle: 'Confirm',
+      vigem: {
+        title: 'ViGEm Bus Driver',
+        hint: 'Base driver for Xbox 360 / DualShock 4 virtual gamepads.',
+        confirmInstall: 'This will install the ViGEm bus driver (admin rights may be required). Continue?',
+        confirmUninstall: 'This will uninstall the ViGEm bus driver; Xbox 360 / DualShock 4 emulation will stop working. Continue?',
+      },
+      vmouse: {
+        title: 'Virtual Mouse Driver',
+        hint: 'Virtual HID driver for high-precision relative mouse movement; falls back to SendInput when unavailable.',
+        enableShort: 'Enable',
+        confirmInstall: 'This will install the virtual mouse driver (admin rights may be required). Continue?',
+        confirmUninstall: 'This will uninstall the virtual mouse driver; streaming mouse input will fall back to SendInput. Continue?',
+      },
+      meta: {
+        title: 'ControllerMeta Test',
+        hint: 'Check device detection, input, polling, motion and basic rumble.',
+        launch: 'Launch test tool',
       },
     },
   },
