+++ src/renderer/desktop/i18n/zh.js
@@ -960,7 +960,9 @@ export const zh = {
     virtualDisplay: '基地显示器',
     webStream: 'Web 串流',
     aiAssistant: '米塔',
+    deviceHub: '设备中心',
     controllers: '控制器',
+    controllersHub: '控制器中心',
     dualSense: 'DualSense 模拟',
     uninstallVdd: '卸载 VDD',
     restartGpu: '重启显卡驱动',
@@ -981,6 +983,118 @@ export const zh = {
     eyebrow: 'MANAGE / CONTROLLERS',
     title: '控制器',
     intro: '管理 Sunshine 的虚拟控制器。',
+  },
+
+  deviceHub: {
+    eyebrow: 'MANAGE / DEVICE HUB',
+    title: '设备中心',
+    intro: '管理虚拟设备、本机组件与未来 USB 透传边界；串流参数仍在 WebUI 中配置。',
+    refresh: '刷新',
+    available: '可用',
+    unavailable: '不可用',
+    needsSetup: '需要配置',
+    experimental: '实验性',
+    planned: '规划中',
+    unknown: '未知',
+    statusUnavailable: '无法读取设备状态，请确认 Sunshine 正在运行且版本匹配。',
+    controllerAdvanced: '高级控制器行为（DS4 / 体感 / DSU）',
+    tabs: {
+      overview: '概览',
+      controllers: '控制器',
+      microphone: '麦克风',
+      usb: 'USB 透传',
+      components: '组件与诊断',
+    },
+    runtime: {
+      title: 'DEVICE RUNTIME',
+      ready: '设备运行基础已就绪',
+      needsAttention: '设备运行基础需要处理',
+      host: 'Virtual Device Host',
+      transport: 'USB/IP Transport',
+      owner: '运行所有者',
+    },
+    overview: {
+      controllers: '虚拟控制器',
+      controllersHint: '选择游戏识别的手柄类型，并管理可选的 DualSense 组件。',
+      manageControllers: '管理控制器',
+      microphone: '虚拟麦克风',
+      microphoneMissing: '签名的 Virtual Device Host 组件不可用。',
+      microphoneCapturing: 'Windows 应用当前正在捕获虚拟端点。',
+      microphoneIdle: '端点已就绪，正在等待应用打开。',
+      microphoneWaiting: '组件已就绪，将在需要时创建设备端点。',
+      statusCapturing: '捕获中',
+      statusReady: '已就绪',
+      statusWaiting: '待创建',
+      statusUnavailable: '不可用',
+      inspectMicrophone: '查看状态与测试',
+      usb: 'USB 设备透传',
+      usbHint: '未来客户端设备将显式授权，并且只绑定一个串流会话。',
+      inspectUsb: '查看规划边界',
+    },
+    microphone: {
+      title: 'VIRTUAL MICROPHONE',
+      experimentalTitle: '实验性 USB/IP 麦克风',
+      experimentalHint: '此路径使用签名的 Virtual Device Host 和固定版本 USB/IP 传输，不经过 VB-Cable 回环。',
+      test: '测试麦克风路径',
+      testSuccess: '测试音已注入虚拟麦克风。',
+      testFailed: '麦克风路径测试未完成。',
+      configuredBackend: '配置后端',
+      activeBackend: '当前后端',
+      component: '宿主组件',
+      endpoint: 'Windows 端点',
+      capture: '应用捕获',
+      buffer: '主机缓冲',
+      health: '传输状态',
+      healthOk: '正常',
+      healthAttention: '需要处理',
+      endpointReady: '已就绪',
+      endpointNotCreated: '未创建',
+      captureActive: '正在捕获',
+      captureIdle: '等待应用',
+      diagnostics: '查看累计诊断数据',
+      underruns: '缓冲欠载（累计）',
+      droppedFrames: '未消费帧（累计）',
+      submitErrors: '提交错误（累计）',
+      backends: {
+        usbip_experimental: 'USB/IP 虚拟麦克风',
+        auto: '自动选择',
+        vb_cable: 'VB-Cable 兼容模式',
+        disabled: '已禁用',
+      },
+      fallback: '当前正在使用后端回退',
+      errorHint: '请检查 Sunshine 日志；若错误持续，请修复组件。',
+      webuiHint: '麦克风启停与后端选择仍在 WebUI 的音频/视频设置中完成。',
+      openWebui: '打开 WebUI 音频设置',
+      openFailed: '无法打开 Sunshine WebUI。',
+      states: {
+        absent: '未创建',
+        starting: '正在启动',
+        ready: '已就绪',
+        idle: '已就绪',
+        remote_active: '正在接收远端音频',
+        host_capturing: '应用正在捕获',
+        faulted: '发生故障',
+        unsupported: '当前平台不支持',
+      },
+    },
+    usb: {
+      title: '通用 USB 透传',
+      notAvailable: '当前版本尚未提供',
+      hint: '这里先固定产品边界；在协议与客户端能力落地前，不展示虚假设备或不可用操作。',
+      direction: '方向',
+      clientToHost: 'Moonlight 客户端 → Sunshine 主机',
+      ownership: '所有权',
+      sessionBound: '绑定一个已认证串流',
+      release: '断开行为',
+      autoRelease: '自动归还客户端',
+      boundaryTitle: '与虚拟设备创建保持独立',
+      boundaryHint: 'Virtual Device Host 创建 Sunshine 自有设备；未来 USB Passthrough Manager 负责枚举、授权、声明和释放外部物理设备。',
+    },
+    components: {
+      runtimeTitle: '共享运行组件',
+      hostHint: '创建 Sunshine 自有的 DualSense 与虚拟麦克风设备。',
+      transportHint: '固定版本的系统传输，由复合虚拟设备与未来透传能力共享。',
+    },
   },
 
   dualSense: {
@@ -1122,6 +1236,74 @@ export const zh = {
           fallback: '设备测试未完成。请点击“修复组件”后重试。',
         },
         uninstall: { fallback: 'DualSense 组件未能卸载。请结束串流，并在提示时允许管理员权限后重试。' },
+      },
+    },
+  },
+
+  controllersHub: {
+    eyebrow: 'MANAGE / CONTROLLERS HUB',
+    title: '控制器中心',
+    intro: '选择串流时游戏识别的手柄类型；更多功能收在高级里。',
+    advancedTitle: '高级（DualSense / 体感 / 驱动与工具）',
+    steamNote: 'Steam Controller 支持开发中',
+    emulation: {
+      typeQuestion: '串流时，让游戏识别为哪种手柄？',
+      familyAuto: '自动（推荐）',
+      familyAutoHint: '不确定时选这个，优先采用兼容性最好的输出。',
+      xbox360: 'Xbox 360',
+      xbox360Hint: '兼容大多数 Windows 游戏，适合优先考虑兼容性的场景。',
+      ds4ModeHint: '适合需要 PlayStation 按键布局、触控板或体感的游戏。',
+      ds5Variant: 'DualSense',
+      ds5ModeHint: '提供自适应扳机与 HD Haptics；需要 Virtual Device Host 组件。',
+      sunshineUnreachable: '无法读取 Sunshine 配置，请确认 Sunshine 正在运行。',
+      retry: '重试',
+      saveFailed: '保存失败，已恢复原设置。',
+      perAppTitle: '每个游戏的单独设置',
+      perAppHint: '为单个游戏单独指定手柄类型，优先于上面的全局选择。',
+      noApps: '暂无应用，请先在 Sunshine WebUI 中添加。',
+      appLoadFailed: '应用列表加载失败。',
+      appSaveConflict: '应用配置已被其他页面修改，已重新加载列表，请重试。',
+      appSaved: '应用手柄类型已保存。',
+      appSaveFailed: '应用手柄类型保存失败，已恢复原设置。',
+      followGlobal: '跟随全局',
+      ds4Title: 'DualShock 4 细节',
+      ds4Hint: '仅影响识别为 DualShock 4 时的输入行为。',
+      backAsTouchpadClick: 'BACK 键映射为触控板点击',
+      motionAsDs4: '体感数据以 DS4 方式注入',
+      touchpadAsDs4: '触控板以 DS4 方式注入',
+      dsuTitle: 'DSU 体感服务器',
+      dsuHint: '开启后，体感辅助工具可通过 UDP 读取手柄姿态数据。',
+      dsuEnable: '启用体感服务器',
+      dsuPort: '端口',
+    },
+    hardware: {
+      comingSoonTitle: '即将支持',
+      comingSoonHint: '该控制器类型已规划，专属设置正在开发中。',
+    },
+    peripherals: {
+      title: '周边工具',
+      installed: '已安装',
+      notInstalled: '未安装',
+      install: '安装',
+      uninstall: '卸载',
+      confirmTitle: '确认操作',
+      vigem: {
+        title: 'ViGEm 总线驱动',
+        hint: 'Xbox 360 / DualShock 4 虚拟手柄的基础驱动。',
+        confirmInstall: '将安装 ViGEm 总线驱动（可能需要管理员权限）。继续？',
+        confirmUninstall: '将卸载 ViGEm 总线驱动，Xbox 360 / DualShock 4 仿真将不可用。继续？',
+      },
+      vmouse: {
+        title: '虚拟鼠标驱动',
+        hint: '高精度相对鼠标移动的虚拟 HID 驱动，不可用时回退 SendInput。',
+        enableShort: '启用',
+        confirmInstall: '将安装虚拟鼠标驱动（可能需要管理员权限）。继续？',
+        confirmUninstall: '将卸载虚拟鼠标驱动，串流鼠标将回退 SendInput。继续？',
+      },
+      meta: {
+        title: 'ControllerMeta 手柄测试',
+        hint: '检测设备识别、输入、轮询、体感与基础振动。',
+        launch: '启动测试工具',
       },
     },
   },
